+++ CHANGES.rst
@@ -6,6 +6,10 @@ Async PRAW follows `semantic versioning <https://semver.org/>`_.
 Unreleased
 ----------
 
+**Fixed**
+
+- XML parsing error when media uploads fail.
+
 7.7.1 (2023/07/11)
 ------------------