+++ CHANGES.rst
@@ -34,12 +34,16 @@ Unreleased
 
 - XML parsing error when media uploads fail.
 
+**Changed**
+
+- Drop support for Python 3.7, which is end-of-life on 2023-06-27.
+
 7.7.1 (2023/07/11)
 ------------------
 
 **Changed**
 
-- Drop asyncio_extras dependency, use contextlib.asynccontextmanager instead.
+- Drop ``asyncio_extras`` dependency, use ``contextlib.asynccontextmanager`` instead.
 
 **Fixed**