+++ CHANGES.rst
@@ -33,6 +33,7 @@ Unreleased
 - :func:`.stream_generator` now accepts the ``continue_after_id`` parameter, which
   starts the stream after a given item ID.
 - Support for new share URL format created from Reddit's mobile apps.
+- Support delayed session creation in asyncprawcore 2.5.0+.
 
 **Changed**