+++ CHANGES.rst
@@ -3,6 +3,9 @@ Change Log
 
 Async PRAW follows `semantic versioning <https://semver.org/>`_.
 
+Unreleased
+----------
+
 7.8.0 (2024/10/20)
 ------------------