+++ CHANGES.rst
@@ -6,6 +6,9 @@ Async PRAW follows `semantic versioning <https://semver.org/>`_.
 Unreleased
 ----------
 
+7.8.1 (2024/12/21)
+------------------
+
 7.8.0 (2024/10/20)
 ------------------